Removed task.go
