+++ backend-docker/pier/job.go
@@ -1,6 +1,7 @@
 package pier
 
 import (
+	"bytes"
 	"github.com/EUDAT-GEF/GEF/backend-docker/pier/internal/dckr"
 	"sort"
 	"sync"
@@ -16,6 +17,7 @@ type Job struct {
 	State        *JobState
 	InputVolume  VolumeID
 	OutputVolume VolumeID
+	Tasks        []TaskStatus
 }
 
 // JobState exported
@@ -28,6 +30,19 @@ type JobState struct {
 type JobID string
 
 type jobArray []Job
+
+// TaskStatus exported
+type TaskStatus struct {
+	Name          string
+	Error         error
+	ExitCode      int
+	ConsoleOutput *bytes.Buffer
+}
+
+type LatestOutput struct {
+	Name          string
+	ConsoleOutput string
+}
 
 func (jl jobArray) Len() int {
 	return len(jl)
@@ -105,5 +120,20 @@ func (jobList *JobList) setOutputVolume(jobID JobID, outputVolume VolumeID) {
 	defer jobList.Unlock()
 	job := jobList.cache[jobID]
 	job.OutputVolume = outputVolume
+	jobList.cache[jobID] = job
+}
+
+func (jobList *JobList) addTask(jobID JobID, taskName string, taskError error, taskExitCode int, taskConsoleOutput *bytes.Buffer) {
+	jobList.Lock()
+	defer jobList.Unlock()
+	job := jobList.cache[jobID]
+
+	var newTask TaskStatus
+	newTask.Name = taskName
+	newTask.Error = taskError
+	newTask.ExitCode = taskExitCode
+	newTask.ConsoleOutput = taskConsoleOutput
+	job.Tasks = append(job.Tasks, newTask)
+
 	jobList.cache[jobID] = job
 }
